# docs/api/openApiSpecs.json
{
  "info": {
    "title": "E-learning Service",
    "version": "0.0.1",
    "description": "E-learning API documentation"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "paths": {
...
    "/lessons/{lessonId}": {
      "get": {
        "tags": [
          "lessons"
        ],
        "summary": "To get list of all lessons.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "lessonId",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/lessons/res.json#GetLesson"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "lessons"
        ],
        "summary": "To update lessons.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "lessonId",
            "type": "integer"
          },
          {
            "in": "body",
            "name": "lesson",
            "description": "lessons data",
            "schema": {
              "$ref": "components/lessons/req.json#UpdateLesson"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "lessons"
        ],
        "summary": "To delete lessons.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "lessonId",
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
    "/modules/{moduleId}": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "To get list of all modules.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "moduleId",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/module/res.json#GetModule"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "modules"
        ],
        "summary": "To update modules.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "moduleId",
            "type": "integer"
          },
          {
            "in": "body",
            "name": "module",
            "description": "module data",
            "schema": {
              "$ref": "components/module/req.json#UpdateModule"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "modules"
        ],
        "summary": "To delete specific module.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "moduleId",
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
    "/modules/{moduleId}/lessons": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "To get list of all modules.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "moduleId",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation.",
            "schema": {
              "$ref": "components/module/res.json#GetModuleLesson"
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "modules"
        ],
        "summary": "To update modules.",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "moduleId",
            "type": "integer"
          },
          {
            "in": "body",
            "name": "module",
            "description": "module data",
            "schema": {
              "$ref": "components/module/req.json#CreateModuleLesson"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation."
          },
          "400": {
            "description": "Bad Request - validation error",
            "schema": {
              "$ref": "components/errorContracts.json#/ValidationErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "403": {
            "description": "forbidden access.",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "components/errorContracts.json#/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}